feat: extract plugin and theme resource selection logic into a reusable ResourceSelector component
