+++ src/components/index.js
@@ -7,3 +7,4 @@ export { default as ThemeSkeleton } from './skeleton/theme-skeleton';
 export { default as StepWrapper } from './step-wrapper';
 export { default as KeyValueEditor } from './key-value-editor';
 export { default as TreeEditor } from './tree-editor';
+export { default as ResourceSelector } from './resource-selector';
